+++ amethyst_audio/src/bundle.rs
@@ -1,10 +1,8 @@
 //! ECS audio bundles
 
 use amethyst_assets::Processor;
-use amethyst_core::{
-    bundle::{Result, SystemBundle},
-    specs::prelude::DispatcherBuilder,
-};
+use amethyst_core::{bundle::SystemBundle, specs::prelude::DispatcherBuilder};
+use amethyst_error::Error;
 
 use crate::source::*;
 
@@ -17,7 +15,7 @@ use crate::source::*;
 pub struct AudioBundle;
 
 impl<'a, 'b> SystemBundle<'a, 'b> for AudioBundle {
-    fn build(self, builder: &mut DispatcherBuilder<'a, 'b>) -> Result<()> {
+    fn build(self, builder: &mut DispatcherBuilder<'a, 'b>) -> Result<(), Error> {
         builder.add(Processor::<Source>::new(), "source_processor", &[]);
         Ok(())
     }