+++ amethyst_network/src/error.rs
@@ -1,74 +1,55 @@
 //! Module containing error handling logic.
 
-use std::{
-    fmt::{self, Display, Formatter},
-    io,
-    sync::mpsc,
-};
+use err_derive::Error;
+use std::{io, sync::mpsc};
 
 use crate::server::ServerSocketEvent;
 
 /// The `amethyst_network` result type.
-pub type Result<T> = std::result::Result<T, ErrorKind>;
+pub type Result<T> = std::result::Result<T, Error>;
 
 /// Wrapper for all errors who could occur in `amethyst_network`.
-#[derive(Debug)]
-pub enum ErrorKind {
+#[derive(Debug, Error)]
+pub enum Error {
     /// Error that could occur on the UDP-socket
+    // NB: NetworkError does not implement std::error::Error and cannot be used as a cause.
+    // But in order to get _some_ useful diagnostics out of it we format it in the message.
+    #[error(display = "UDP-error occurred: {}", _0)]
     UdpError(laminar::error::NetworkError),
     /// Error that could occur whit IO.
-    IoError(io::Error),
+    #[error(display = "IO-error occurred")]
+    IoError(#[cause] io::Error),
     /// Error that could occur when serializing whit `bincode`
-    SerializeError(bincode::ErrorKind),
+    #[error(display = "Serialization error occurred")]
+    SerializeError(#[cause] bincode::Error),
     /// Error that could occur when sending an `ServerSocketEvent` to some channel.
-    ChannelSendError(mpsc::SendError<ServerSocketEvent>),
+    #[error(display = "Channel send error occurred")]
+    ChannelSendError(#[cause] mpsc::SendError<ServerSocketEvent>),
+    #[error(display = "Some error has occurred")]
     #[doc(hidden)]
     __Nonexhaustive,
 }
 
-impl std::error::Error for ErrorKind {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match *self {
-            ErrorKind::IoError(ref e) => Some(e),
-            ErrorKind::ChannelSendError(ref e) => Some(e),
-            ErrorKind::SerializeError(ref e) => Some(e),
-            _ => None,
-        }
+impl From<io::Error> for Error {
+    fn from(e: io::Error) -> Error {
+        Error::IoError(e)
     }
 }
 
-impl Display for ErrorKind {
-    fn fmt(&self, fmt: &mut Formatter<'_>) -> fmt::Result {
-        match *self {
-            ErrorKind::UdpError(_) => write!(fmt, "UDP-error occurred"),
-            ErrorKind::IoError(_) => write!(fmt, "IO-error occurred"),
-            ErrorKind::ChannelSendError(_) => write!(fmt, "Channel send error occurred"),
-            ErrorKind::SerializeError(_) => write!(fmt, "Channel send error occurred"),
-            _ => write!(fmt, "Some error has occurred"),
-        }
+impl From<mpsc::SendError<ServerSocketEvent>> for Error {
+    fn from(e: mpsc::SendError<ServerSocketEvent>) -> Error {
+        Error::ChannelSendError(e)
     }
 }
 
-impl From<io::Error> for ErrorKind {
-    fn from(e: io::Error) -> ErrorKind {
-        ErrorKind::IoError(e)
+impl From<laminar::error::NetworkError> for Error {
+    fn from(e: laminar::error::NetworkError) -> Error {
+        Error::UdpError(e)
     }
 }
 
-impl From<mpsc::SendError<ServerSocketEvent>> for ErrorKind {
-    fn from(e: mpsc::SendError<ServerSocketEvent>) -> ErrorKind {
-        ErrorKind::ChannelSendError(e)
-    }
-}
-
-impl From<laminar::error::NetworkError> for ErrorKind {
-    fn from(e: laminar::error::NetworkError) -> ErrorKind {
-        ErrorKind::UdpError(e)
-    }
-}
-
-impl From<Box<bincode::ErrorKind>> for ErrorKind {
-    fn from(e: Box<bincode::ErrorKind>) -> ErrorKind {
-        ErrorKind::SerializeError(*e)
+impl From<bincode::Error> for Error {
+    fn from(e: bincode::Error) -> Error {
+        Error::SerializeError(e)
     }
 }