+++ amethyst_utils/src/auto_fov.rs
@@ -1,10 +1,11 @@
 //! Utility to adjust the aspect ratio of cameras automatically
 
-use amethyst_assets::{PrefabData, PrefabError};
+use amethyst_assets::PrefabData;
 use amethyst_core::specs::{
     Component, Entity, HashMapStorage, Join, ReadExpect, ReadStorage, System, WriteStorage,
 };
 use amethyst_derive::PrefabData;
+use amethyst_error::Error;
 use amethyst_renderer::{Camera, ScreenDimensions};
 
 use serde::{Deserialize, Serialize};