+++ amethyst_utils/src/removal.rs
@@ -2,11 +2,12 @@
 
 use std::fmt::Debug;
 
-use amethyst_assets::{PrefabData, PrefabError};
+use amethyst_assets::PrefabData;
 use amethyst_core::specs::{
     world::EntitiesRes, Component, DenseVecStorage, Entity, Join, ReadStorage, WriteStorage,
 };
 use amethyst_derive::PrefabData;
+use amethyst_error::Error;
 
 use log::error;
 use serde::{Deserialize, Serialize};