+++ docs/CHANGELOG.md
@@ -27,6 +27,7 @@ it is attached to. ([#1282])
 * Remove unnecessary texture coordinates offset in `Sprite::from_pixel_values` ([#1267])
 * Changed `ActiveCamera` to have the `Option` inside. ([#1280])
 * `AudioBundle::new()` no longer exists, as `AudioBundle` is now a unit type. It also no longer initializes the `DjSystem` ([#1356])
+* Convert everything to use err-derive and amethyst_error ([#1365])
 
 ### Removed
 
@@ -50,6 +51,7 @@ it is attached to. ([#1282])
 [#1328]: https://github.com/amethyst/amethyst/pull/1328
 [#1356]: https://github.com/amethyst/amethyst/pull/1356
 [#1363]: https://github.com/amethyst/amethyst/pull/1363
+[#1365]: https://github.com/amethyst/amethyst/pull/1365
 [#1371]: https://github.com/amethyst/amethyst/pull/1371
 
 ## [0.10.0] - 2018-12